Refactor PipelineProgress
- Extract time estimation logic into useTimeEstimation hook
- Extract file tracking logic into useFileTracking hook
- Extract progress calculation logic into useProgressCalculation hook
- Extract dataset item rendering into DatasetProgressItem component
